Fix for infinite loop in inotify buffer parser.
If an unexpected directory is seen then the parser would
log the directory information and then continue without
incrementing the cursor. This means that the parser will
sit in an infinite loop printing the same debug line if
a non-standard directory name is seen.

Also the parser would not increment the cursor if the event->len
was zero. This edge case is unlikely to happen, but is also
mitigated by this code change.