Add support for clean shutdown.
The previous version called abort() to shutdown the process. The
disadvantage of this is that it can result in partially written
files. This change modifies the code so that the TERM signal will
result in a graceful shutdown that should not lead to partially
written files.

Changes are as follows:

- Allow null pointers to be pushed onto the queue. The file copy
  threads will stop if they read the null.
- Use a select() call with timeout in the inotify threads to allow
  them to wake up periodically to check the run status.
- Change to use sigaction() rather than the deprecated signal() api
  for signal handling.